backend/requirements.txt
fastapi
uvicorn         # ASGI server to run FastAPI
requests       # For making HTTP requests to other APIs
python-dotenv    # Load environment variables from .env
SQLAlchemy  # ORM of python classes and DB
psycopg2-binary #  Driver for PostgreSQL
databases[postgresql]